refactor(transcription): migrate from whisper-rs to transcribe-rs
Replace whisper-rs with transcribe-rs library for unified GPU acceleration across platforms. This consolidates the previous platform-specific GPU features (Metal on macOS, Vulkan on Windows/Linux) into a single transcription engine.

Key changes:
- Rename whisper_cpp module to transcription for broader scope
- Replace WhisperCppError with TranscriptionError for consistency
- Update Cargo.toml to use transcribe-rs instead of platform-specific whisper-rs features
- Refactor transcribe_with_whisper_cpp to transcribe_audio with new library
- Maintain same audio conversion and processing logic

The transcribe-rs library provides automatic GPU detection and acceleration, simplifying the build configuration while maintaining performance benefits.
