+++ apps/whispering/src-tauri/Cargo.toml
@@ -47,6 +47,7 @@ lazy_static = "1.4"
 tempfile = "3.8"
 tauri-plugin-macos-permissions = "2.3.0"
 fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs" }
+transcribe-rs = "0.1.0"
 
 [target.'cfg(unix)'.dependencies]
 nix = { version = "0.29", features = ["signal"] }
@@ -57,21 +58,11 @@ windows-sys = { version = "0.59", features = ["Win32_Foundation", "Win32_System_
 [target.'cfg(target_os = "macos")'.dependencies]
 accessibility-sys =  "0.1.3"
 core-foundation-sys =  "0.8.7"
-# macOS: GPU acceleration using Metal
-whisper-rs = { version = "0.15.0", features = ["metal"] }
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-global-shortcut = "2"
 tauri-plugin-single-instance = "2"
 tauri-plugin-updater = "2"
-
-[target.'cfg(target_os = "windows")'.dependencies]
-# Windows: GPU acceleration using Vulkan
-whisper-rs = { version = "0.15.0", features = ["vulkan"] }
-
-[target.'cfg(all(unix, not(target_os = "macos")))'.dependencies]
-# Linux: GPU acceleration using Vulkan
-whisper-rs = { version = "0.15.0", features = ["vulkan"] }
 
 [profile.dev]
 incremental = true # Compile your binary in smaller steps.
