+++ apps/whispering/src-tauri/src/transcription/error.rs
@@ -1,0 +1,28 @@
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+
+#[derive(Error, Debug, Serialize, Deserialize)]
+pub enum TranscriptionError {
+    #[error("Audio read error: {message}")]
+    AudioReadError { message: String },
+
+    #[error("GPU error: {message}")]
+    GpuError { message: String },
+
+    #[error("Model load error: {message}")]
+    ModelLoadError { message: String },
+
+    #[error("Transcription error: {message}")]
+    TranscriptionError { message: String },
+}
+
+impl TranscriptionError {
+    pub fn name(&self) -> &'static str {
+        match self {
+            TranscriptionError::AudioReadError { .. } => "AudioReadError",
+            TranscriptionError::GpuError { .. } => "GpuError",
+            TranscriptionError::ModelLoadError { .. } => "ModelLoadError",
+            TranscriptionError::TranscriptionError { .. } => "TranscriptionError",
+        }
+    }
+}
